Move all compatibility specs to a single folder
