+++ spec/compatibility/rubocop/config_spec.rb
@@ -4,7 +4,7 @@ require "open3"
 
 RSpec.describe 'Rubocop default config' do
   def run_rubocop(path)
-    md_path = File.expand_path("../../lib/anycable/rails/compatibility/cops.rb", __dir__)
+    md_path = File.expand_path("../../../lib/anycable/rails/compatibility/cops.rb", __dir__)
     output, _status = Open3.capture2(
       "bundle exec rubocop -d -r #{md_path} #{path}",
       chdir: File.join(__dir__, "fixtures")
