+++ spec/compatibility/rubocop/cops/instance_vars_spec.rb
