+++ spec/compatibility/rubocop/cops/periodical_timers_spec.rb
