+++ spec/compatibility/rubocop/cops/remote_disconnect_spec.rb
