+++ spec/compatibility/rubocop/cops/stream_callbacks_spec.rb
