+++ spec/compatibility/rubocop/fixtures/app/channels/bad_channel.rb
