+++ spec/compatibility/rubocop/fixtures/app/controllers/good_controller.rb
